feat(embeddings): introduce factory functions for embedding providers
- Added a new `factory.rs` file containing functions to create embedding providers based on specified configurations, supporting Ollama, OpenAI, custom endpoints, and a no-op provider.
- Implemented error handling for unrecognized provider names to ensure immediate feedback on configuration issues.
- Updated `mod.rs` to include the new factory module, enhancing code organization and maintainability.
- Introduced a `provider_trait.rs` file defining the `EmbeddingProvider` trait, establishing a consistent interface for embedding implementations.
