+++ src/openhuman/embeddings/provider_trait.rs
@@ -1,0 +1,24 @@
+//! Interface for embedding providers that convert text into numerical vectors.
+
+use async_trait::async_trait;
+
+/// Interface for embedding providers that convert text into numerical vectors.
+#[async_trait]
+pub trait EmbeddingProvider: Send + Sync {
+    /// Returns the name of the provider (e.g., "ollama", "openai").
+    fn name(&self) -> &str;
+
+    /// Returns the number of dimensions in the generated embeddings.
+    fn dimensions(&self) -> usize;
+
+    /// Generates embeddings for a batch of strings.
+    async fn embed(&self, texts: &[&str]) -> anyhow::Result<Vec<Vec<f32>>>;
+
+    /// Generates an embedding for a single string.
+    async fn embed_one(&self, text: &str) -> anyhow::Result<Vec<f32>> {
+        let mut results = self.embed(&[text]).await?;
+        results
+            .pop()
+            .ok_or_else(|| anyhow::anyhow!("Empty embedding result"))
+    }
+}
